requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Azure Document Intelligence SDK
azure-ai-formrecognizer==3.3.2

# OpenAI API
openai==1.30.1

# Chroma vector store
chromadb==0.4.24

# For file handling and async support
python-multipart==0.0.9

# Optional: for local development and testing
httpx==0.27.0
